refactor(cblmariner) split types
